+++ src/greentest/test__pool.py
@@ -3,6 +3,7 @@ import gevent
 from gevent import pool
 from gevent.event import Event
 from gevent.queue import Queue
+from gevent.timeout import Timeout
 import greentest
 import random
 from greentest import ExpectedException
@@ -216,6 +217,51 @@ class PoolBasicTests(greentest.TestCase):
             self.assertEqual(len(p), 0)
         finally:
             g.kill()
+
+    @greentest.ignores_leakcheck
+    def test_add_method_non_blocking(self):
+        p = self.klass(size=1)
+        first = gevent.spawn(gevent.sleep, 1000)
+        try:
+            second = gevent.spawn(gevent.sleep, 1000)
+            try:
+                p.add(first)
+                with self.assertRaises(pool.PoolFull):
+                    p.add(second, blocking=False)
+            finally:
+                second.kill()
+        finally:
+            first.kill()
+
+    @greentest.ignores_leakcheck
+    def test_add_method_timeout(self):
+        p = self.klass(size=1)
+        first = gevent.spawn(gevent.sleep, 1000)
+        try:
+            second = gevent.spawn(gevent.sleep, 1000)
+            try:
+                p.add(first)
+                with self.assertRaises(pool.PoolFull):
+                    p.add(second, timeout=0.100)
+            finally:
+                second.kill()
+        finally:
+            first.kill()
+
+    @greentest.ignores_leakcheck
+    def test_start_method_timeout(self):
+        p = self.klass(size=1)
+        first = gevent.spawn(gevent.sleep, 1000)
+        try:
+            second = gevent.Greenlet(gevent.sleep, 1000)
+            try:
+                p.add(first)
+                with self.assertRaises(pool.PoolFull):
+                    p.start(second, timeout=0.100)
+            finally:
+                second.kill()
+        finally:
+            first.kill()
 
     def test_apply(self):
         p = self.klass()